+++ src/com/feup/jhotsketch/model/RoundedRectangleModel.java
@@ -43,7 +43,7 @@ public class RoundedRectangleModel extends ShapeModel{
 		handles.add(new Handle(this, 0, radiusy, "ROUNDRADIUSY", "RHOMBUS", SWT.COLOR_DARK_GREEN));
 	}
 	
-	public void moveHandle(double rx, double ry, Handle handle){
+	public void moveHandle(int rx, int ry, Handle handle){
 		super.moveHandle(rx, ry, handle);
 		if (handle.getId().equals("ROUNDRADIUSX")) radiusx += rx;
 		if (handle.getId().equals("ROUNDRADIUSY")) radiusy += ry;
